requirements.txt
selenium~=4.9.0
webdriver-manager
pytest~=7.3.1
allure-pytest
requests
pytest-xdist
PyYAML~=6.0